【修改描述】日志缩减&flow资源预取边界值修正 【DTS/CR/AR单号】DTS2024051707186 【合入版本】Open Computing Kit for Kunpeng 24.0.0
Match-id-883890690a70611f3b28de94e4afda47eb1b4d15

+++ src/cache/read/rcache_chunk.h
@@ -121,7 +121,7 @@ private:
     Key mKey;
     RCacheValue mValue;
     uint32_t aTime;
-    uint8_t hitCount{ 0 };
+    uint16_t hitCount{ 0 };
     RCacheTierType tierType;
     MqType mMqType;
     uint64_t mState{ 0 };
